refactoring some of the frontend code continued
