+++ frontend/src/components/OcpPerformanceHeader.js
@@ -1,10 +1,8 @@
 import React from 'react';
 import { Title, TitleSizes } from '@patternfly/react-core';
-import SimpleTabs from './PerformanceTabs';
 
 
-const OcpPerformanceHeader = (props) => {
-  const versions = props.data
+const OcpPerformanceHeader = () => {
   return (
     <> 
         <img className="OcpPerformanceTable-logo" src="logo.png" alt="Openshift Logo"/>
