+++ frontend/src/components/PlatformTabs.js
@@ -1,30 +1,11 @@
 import ReactDOM from 'react-dom';
-// import "@patternfly/react-core/dist/styles/base.css";
-// import './fonts.css';
-
 import React, { useState, useEffect } from 'react';
 import { Tabs, Tab, TabTitleText, Checkbox } from '@patternfly/react-core';
 
-import WithLabels from './WithLabels';
-import VersionList from './VersionList';
+import PerformanceResults from './PerformanceResults';
+import VersionList from './PerformanceResults';
 
-
-// export default function SimpleTabs(props) {
-//   const [activeTabKey, setActiveKey] = useState([]);
-
-//   return (
-//     <Tabs activeKey={activeTabKey} onSelect={setActiveKey(index)}>
-//       {props.data.map((tab, index) => (      
-//         <Tab eventKey={index} title={<TabTitleText>tab.title</TabTitleText>}>
-//           <WithLabels data={tab.versions} />
-//           <VersionList data={tab.versions} />
-//         </Tab>
-//       ))}
-//     </Tabs>
-//   );
-// }
-
-class SimpleTabs extends React.Component {
+class PlatformTabs extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -46,7 +27,7 @@ class SimpleTabs extends React.Component {
       <Tabs activeKey={activeTabKey} onSelect={this.handleTabClick}>
       {this.props.data.map((tab, index) => (         
         <Tab eventKey={index} title={<TabTitleText>{tab.title}</TabTitleText>}> 
-          <WithLabels data={tab.data} />
+          <PerformanceResults data={tab.data} />
           <VersionList data={tab.data} />
          </Tab>
        ))
@@ -56,4 +37,4 @@ class SimpleTabs extends React.Component {
     );
   }
 }
-export default SimpleTabs;
+export default PlatformTabs;
